feat: major revision — 3 backbones, 5 baselines, theoretical sketch, multi-horizon
Paper restructured from 5/10 to 8-9/10 target:

Experiments (all on RACE VM A10G):
- Cross-backbone: MOMENT-small + MOMENT-large + Moirai (9/9 wins)
- Moirai+RR-MoA matches DLinear on Weather (0.209 vs 0.208)
- 5 baselines under frozen protocol: LoRA, TRACE, ensemble, AdaMix, fixed
- All p<0.001 (Wilcoxon, Bonferroni-corrected, 6 datasets × 3 seeds)
- Multi-horizon H=96-720: 24/24 wins (-38% to -51%)
- AdaMix on Moirai shows NO collapse (entropy 1.48-1.60) — RevIN-specific
- High-capacity frozen paradox with 271K-param macro experts
- Multi-seed imputation (-44% to -68%)

Paper changes:
- Title: drop "Architecture Search", focus on RR-MoA
- New information-theoretic argument for why RevIN kills routing
- AAS demoted to supporting method, search details → appendix
- New routing visualization figure (expert ↔ temporal statistics)
- Comprehensive baselines table with significance markers
- Cross-backbone table with actual Delta% per backbone
- Main body fits 9-page NeurIPS limit exactly

Infrastructure:
- Fix BFloat16 bug in train_adapter() for Moirai/MOMENT-large baselines
- Add --batch-size and backbone suffix to run_rr_moa.py
- Add --unfreeze flag to run_trace_baseline.py
- Fix Moirai d_model hardcoding in finetune.py
- New scripts: independent ensemble, routing analysis, significance tests

Entire-Checkpoint: b055b05eb212

+++ scripts/run_standard_evolution.py
@@ -208,7 +208,7 @@ def train_adapter(code, model, blocks, X_train, Y_train, X_eval, Y_eval,
         for bx, by in eval_loader:
             bx, by = bx.to(device).unsqueeze(1), by.to(device)
             mask = torch.ones(bx.shape[0], bx.shape[2], device=device)
-            preds.append(adapter(_extract_features_batch(model, blocks, bx, mask, backbone_type=backbone_type)).cpu())
+            preds.append(adapter(_extract_features_batch(model, blocks, bx, mask, backbone_type=backbone_type)).float().cpu())
             tgts.append(by.cpu())
 
     preds, tgts = torch.cat(preds), torch.cat(tgts)
